requirements.txt
phidata
python-dotenv
duckduckgo-search
openai